moves daos config
