+++ src/config/index.ts
@@ -1,2 +1,5 @@
+export * from './actions';
 export * from './common';
-export * from './daoAdapters';
+export * from './daos';
+export * from './events';
+export * from './types';
